+++ tests/browser/click.test.js
@@ -119,4 +119,21 @@ describe("Click", function() {
         await browser.click("label");
         await browser.assert.text("#value-input", "Label");
     });
+
+    it('Clicks XY coordinates', async() => {
+        await browser.open(configUrls.difficultClick);
+        let clicked = await browser.click(10, 10);
+        await browser.assert.text("#clicker", "click me");
+        assert.strictEqual(clicked, 1);
+        clicked = await browser.click(100, 100);
+        assert.strictEqual(clicked, 1);
+        await browser.assert.text("#clicker", "Clicked!");
+    });
+
+    it('Clicks XY coordinates Throws', async() => {
+        await browser.open(configUrls.difficultClick);
+        await utils.assertThrowsAsync(async() => {
+            await browser.click(4000, 4000);
+        }, `QueryError: No element in position [4000, 4000] found when trying to click.`);
+    });
 });